Reduce code diff to roll back extracting new abstract class
